Move to Searchbar and clear input
